Simplify code and test for SVG text expander
